update word and cache language model, and loss
